refactor: 関数類を外部ファイルに移動
